+++ lib/preprocess.js
@@ -6,7 +6,7 @@ import { validateBlock } from '@web3-storage/car-block-validator'
 import { recursive as exporter } from 'ipfs-unixfs-exporter'
 import createDebug from 'debug'
 import pRetry from 'p-retry'
-import fs from 'node:fs/promises'
+import { appendToRoundBuffer } from './round-buffer.js'
 
 const debug = createDebug('spark:preprocess')
 
@@ -56,8 +56,7 @@ export const preprocess = async ({
   fetchMeasurements,
   recordTelemetry,
   logger,
-  fetchRetries = 14,
-  ROUND_BUFFER_PATH
+  fetchRetries = 14
 }) => {
   const start = Date.now()
   /** @type import('./typings.js').RawMeasurement[] */
@@ -118,10 +117,7 @@ export const preprocess = async ({
   logger.log('Retrieval Success Rate: %s%s (%s of %s)', Math.round(100 * okCount / total), '%', okCount, total)
 
   round.measurements.push(...validMeasurements)
-  await fs.appendFile(
-    ROUND_BUFFER_PATH,
-    validMeasurements.map(m => JSON.stringify(m)).join('\n') + '\n'
-  )
+  await appendToRoundBuffer(validMeasurements)
 
   recordTelemetry('preprocess', point => {
     point.intField('round_index', roundIndex)
